Complete conversion of robn from int to size_t.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@605 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ fisher.c
@@ -60,7 +60,7 @@ double prbf(double x, double df)
     return(status==0 ? q : 1.0);
 }
 
-double fis_get_spamicity(int robn, double invlogsum, double logsum, double *invproduct, double *product)
+double fis_get_spamicity(size_t robn, double invlogsum, double logsum, double *invproduct, double *product)
 {
     double df = 2.0 * robn;
     double _invproduct = prbf(-2.0 * invlogsum, df);
