+++ fisher.h
@@ -8,6 +8,6 @@
 
 extern	rf_method_t rf_fisher_method;
 extern	void	fis_initialize_constants(void);
-extern	double	fis_get_spamicity(int robn, double invlogsum, double logsum, double *invproduct, double *product);
+extern	double	fis_get_spamicity(size_t robn, double invlogsum, double logsum, double *invproduct, double *product);
 
 #endif	/* HAVE_FISHER_H */
